+++ NEWS.md
@@ -18,3 +18,4 @@
 * Added ability to relax map points directly from the javascript viewer
 * Added the function `mapGadget()` for opening up a map object in the GUI
 * Fix error causing 3d arrows to not display
+* Update the way point stress and residual is calculated for the stress and residual error table